+++ src/web/routes/streams.ts
@@ -1,6 +1,5 @@
 import { createLogger } from '../../logger';
 import { Router } from 'express';
-import { randomBytes } from 'crypto';
 import {
   getAllStreamGroups,
   addStreamGroup,
@@ -11,16 +10,11 @@ import {
   removeStreamer,
   removeStreamersByGroup,
   getAllEventSubStreamers,
-  saveEventConfig,
-  clearStreamerToken,
   getTwitchEnabledChannels,
-  EventSubConfig,
 } from '../../db';
 import { csrfProtection } from '../csrf';
-import { requireManager, requireAdmin } from '../middleware';
+import { requireManager } from '../middleware';
 import { restartTwitchMonitor, getLiveStates } from '../../twitchMonitor';
-import { reloadEventSubSubscriptions } from '../../twitchEventSub';
-import { TWITCH_CLIENT_ID, TWITCH_EVENTSUB_REDIRECT_URI } from '../../config';
 import { AccessLevel } from '../../db/users';
 
 const log = createLogger('Web');
@@ -46,7 +40,7 @@ const KNOWN_ERRORS = new Set([
   'eventsub_wrong_account',
 ]);
 
-const ERROR_MESSAGES: Record<string, string> = {
+export const ERROR_MESSAGES: Record<string, string> = {
   missing_fields:                'All required fields must be filled in.',
   invalid_id:                    'Invalid ID — please try again.',
   add_group_failed:              'Failed to add stream group. Please try again.',
@@ -225,89 +219,6 @@ router.post('/streams/streamers/remove', requireManager, csrfProtection, async (
     log.error('Remove streamer error:', err);
     return res.redirect('/admin/streams?error=remove_streamer_failed');
   }
-  res.redirect('/admin/streams');
-});
-
-// ─── EventSub OAuth & config ──────────────────────────────────────────────────
-
-const TWITCH_OAUTH_SCOPE = 'moderator:read:followers channel:read:subscriptions';
-
-router.get('/streams/twitch-oauth/:streamerId', requireAdmin, async (req, res) => {
-  const streamerId = parsePositiveIntId(req.params['streamerId']);
-  if (streamerId === null) return res.redirect('/admin/streams?error=invalid_id');
-
-  const streamers = await getAllEventSubStreamers().catch(() => null);
-  const streamer = streamers?.find((s) => s.id === streamerId);
-  if (!streamer) return res.redirect('/admin/streams?error=invalid_id');
-
-  const botEnabled = await getTwitchEnabledChannels().catch(() => [] as string[]);
-  if (!botEnabled.includes(streamer.name)) {
-    return res.redirect('/admin/streams?error=eventsub_not_bot_enabled');
-  }
-
-  const state = randomBytes(16).toString('hex');
-  req.session.eventsubOAuthState = state;
-  req.session.eventsubStreamerId = streamerId;
-
-  const params = new URLSearchParams({
-    client_id: TWITCH_CLIENT_ID,
-    redirect_uri: TWITCH_EVENTSUB_REDIRECT_URI,
-    response_type: 'code',
-    scope: TWITCH_OAUTH_SCOPE,
-    state,
-    force_verify: 'true',
-  });
-
-  res.redirect(`https://id.twitch.tv/oauth2/authorize?${params.toString()}`);
-});
-
-router.post('/streams/twitch-disconnect/:streamerId', requireAdmin, csrfProtection, async (req, res) => {
-  const streamerId = parsePositiveIntId(req.params.streamerId);
-  if (streamerId === null) return res.redirect('/admin/streams?error=invalid_id');
-
-  try {
-    await clearStreamerToken(streamerId);
-    reloadEventSubSubscriptions();
-  } catch (err) {
-    log.error('EventSub disconnect error:', err);
-    return res.redirect('/admin/streams?error=eventsub_disconnect_failed');
-  }
-  res.redirect('/admin/streams');
-});
-
-// ─── EventSub notification config ────────────────────────────────────────────
-
-router.post('/streams/event-config/:streamerId', requireAdmin, csrfProtection, async (req, res) => {
-  const streamerId = parsePositiveIntId(req.params.streamerId);
-  if (streamerId === null) return res.redirect('/admin/streams?error=invalid_id');
-
-  const botEnabled = await getTwitchEnabledChannels().catch(() => [] as string[]);
-  const streamers = await getAllEventSubStreamers().catch(() => null);
-  const streamer = streamers?.find((s) => s.id === streamerId);
-  if (!streamer || !botEnabled.includes(streamer.name)) {
-    return res.redirect('/admin/streams?error=eventsub_not_bot_enabled');
-  }
-
-  const body = req.body as Record<string, string | undefined>;
-  const config: EventSubConfig = {
-    follow_enabled: body.follow_enabled === 'on',
-    follow_message: (body.follow_message ?? '').trim() || 'Thanks {display_name} for the follow!',
-    sub_enabled: body.sub_enabled === 'on',
-    sub_message: (body.sub_message ?? '').trim() || 'Thanks {display_name} for subscribing! (Tier {tier_name})',
-    resub_message: (body.resub_message ?? '').trim() || 'Thanks {display_name} for {months} months! (Tier {tier_name})',
-    giftsub_message: (body.giftsub_message ?? '').trim() || '{gifter_display} gifted {count} sub(s) to the community!',
-    raid_enabled: body.raid_enabled === 'on',
-    raid_message: (body.raid_message ?? '').trim() || 'Welcome raiders from {from_channel}! Thank you for the {viewers} person raid!',
-  };
-
-  try {
-    await saveEventConfig(streamerId, config);
-    reloadEventSubSubscriptions();
-  } catch (err) {
-    log.error('EventSub config save error:', err);
-    return res.redirect('/admin/streams?error=eventsub_config_failed');
-  }
-
   res.redirect('/admin/streams');
 });
