+++ src/web/server.ts
@@ -13,6 +13,7 @@ const log = createLogger('Web');
 const isProduction = process.env.NODE_ENV === 'production';
 import authRouter from './routes/auth';
 import eventsubCallbackRouter from './routes/eventsubCallback';
+import eventsubAdminRouter from './routes/eventsubAdmin';
 import dashboardRouter from './routes/dashboard';
 import adminRouter from './routes/admin';
 import apiRouter from './routes/api';
@@ -127,6 +128,7 @@ app.use('/', requireAuth, streamdeckKeysRouter);
 app.use('/', requireAuth, sfxRouter);
 app.use('/admin', requireAuth, adminRouter);
 app.use('/admin', requireAuth, streamsRouter);
+app.use('/admin', requireAuth, eventsubAdminRouter);
 app.use('/admin', requireAuth, commandsRouter);
 app.use('/admin', requireAuth, countersRouter);
 app.use('/admin', requireAuth, commandMonitorRouter);
